+++ dolphin/async/src/main/java/edu/snu/cay/dolphin/async/examples/addvector/AddVectorTrainer.java
@@ -17,7 +17,7 @@ package edu.snu.cay.dolphin.async.examples.addvector;
 
 import edu.snu.cay.common.math.linalg.Vector;
 import edu.snu.cay.common.metric.MetricsMsgSender;
-import edu.snu.cay.common.param.Parameters;
+import edu.snu.cay.dolphin.async.DolphinParameters;
 import edu.snu.cay.dolphin.async.EpochInfo;
 import edu.snu.cay.dolphin.async.MiniBatchInfo;
 import edu.snu.cay.dolphin.async.Trainer;
@@ -87,8 +87,8 @@ final class AddVectorTrainer implements Trainer {
 
   @Inject
   private AddVectorTrainer(final ParameterWorker<Integer, Integer, Vector> parameterWorker,
-                           @Parameter(Parameters.MaxNumEpochs.class) final int maxNumEpochs,
-                           @Parameter(Parameters.MiniBatchSize.class) final int miniBatchSize,
+                           @Parameter(DolphinParameters.MaxNumEpochs.class) final int maxNumEpochs,
+                           @Parameter(DolphinParameters.MiniBatchSize.class) final int miniBatchSize,
                            @Parameter(ExampleParameters.DeltaValue.class) final int delta,
                            @Parameter(ExampleParameters.NumKeys.class) final int numberOfKeys,
                            @Parameter(ExampleParameters.NumWorkers.class) final int numberOfWorkers,
